refactor : moved constants.js to config
