+++ client/src/app/routes/ue/Settings.jsx
@@ -7,7 +7,7 @@ import {
 } from "react-router-dom";
 import ValidatedInput from "../../../components/ValidatedInput";
 import ValidatedForm from "../../../components/ValidatedForm";
-import { USE_MOCK_DATA } from "../../../constants";
+import { USE_MOCK_DATA } from "../../../config/constants";
 import axiosInstance from "../../../config/axiosConfig";
 
 export async function action({ request, params }) {
